Add a script that contains function measure_time
Implemented a function measure_time, that accepst two arguments: n(Int) and max_delay(Int)

+++ 0x01-python_async_function/2-measure_runtime.py
@@ -5,13 +5,11 @@
 """
 import time
 import asyncio
-import random
-
 
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-def measure_time(n: int, max_delay: int) -> int:
+def measure_time(n: int, max_delay: int) -> float:
     """
         This function measures the execution time for
         wait_n(n, max_delay)
